Add project root as first item in cdp.ps1 selection
a5034a496bf3ac5785fdb0afc39e2afc84f3f26b `cd`s into the project root first and then shows the
selection so that I can Ctrl+C out of it to get to the project root.

This adds the project root as the first option instead of `cd`ing into the project root, so that the
directory isn't changed if I Ctrl+C out of cdp.ps1.